requirements-docker.txt
# Cloud Run / Docker 전용 — 서버 구동에 필요한 패키지만 포함
# 개발 전용 패키지(gradio, pyaudio, pygame, ddgs 등)는 제외해 이미지 크기 최소화

torch==2.4.1
torchvision==0.19.1
numpy==1.26.4

ultralytics==8.3.2

opencv-python-headless==4.10.0.84
Pillow>=10.0.0

fastapi==0.115.5
uvicorn==0.32.1
python-multipart==0.0.12
python-dotenv>=1.0.0

gTTS==2.5.3
requests>=2.32.0
elevenlabs>=2.0.0

easyocr

psycopg[binary]>=3.1.0
psycopg-pool>=3.1.0
